Use NRI as replacement of CNI
CNI is a stateless specification based on executing binaries.
The shared model is based on chaining, and requires integrations to
write configuration files in disk, with not easy way to detect
incompatibilities.

Kubernetes does not require CNI, is a detal of the controller runtimes,
and this only use it to create an interface in a Pod and return the
assigned IPs.

Since Istio does not require to configure the low level pod networking
and only need to participate in the Pod creation and get the Pod
assigned IPs and its network namespace, it is the perfect candidate to
use NRI, that has been enhanced to exactly solve this integration
problem.

In addition, instead of using the userspace iptables tooling, use the
netlink interface in the kernel to program the redirecting rules and
remove dependencies.

Change-Id: Ide5da6f2308fa61f8ddaec51c4c8d00edd96075c